test3

+++ app/Http/Controllers/CvController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Http\UploadedFile;
+use Auth;
 use App\Cv;
 use App\Experience;
 class CvController extends Controller
@@ -52,8 +53,9 @@ class CvController extends Controller
         }*/
         $url  = $request->imgUpload1->store('img');
         $cv->photo  = $url;
-        $cv->user_id      = Auth::user()->id;
+        $cv->user_id  = Auth::user()->id;
         $cv->save();
+        session()->flash('succes','le CV bien enregistré !!!');
         return redirect('cvs');
     }
 
@@ -65,9 +67,15 @@ class CvController extends Controller
      */
     public function show($id)
     {
-        return view('cv.show',['id' => $id]);
+        $cv =  Cv::find($id);
+        
+        if($cv->user_id  === Auth::user()->id){
+            return view('cv.show',['id' => $id]);
+        }
+        else{
+            return view('cv.admin',['id' => $id]);
+        }
     }
-
     /**
      * Show the form for editing the specified resource.
      *
@@ -93,7 +101,8 @@ class CvController extends Controller
         $cv =  Cv::find($id);
         $cv->title = $request->input('title');
         $cv->presentation = $request->input('text');
-
+        $url  = $request->imgUpload1->store('img');
+        $cv->photo  = $url;
         $cv->save();
         return redirect('cvs');
     }
@@ -107,6 +116,7 @@ class CvController extends Controller
     public function destroy($id)
     {
         $cv =  Cv::find($id);
+        $this->authorize('delete' , $cv);
         $cv->delete();
         return redirect('cvs');
     }
@@ -117,12 +127,12 @@ class CvController extends Controller
         $formations = $cv->formations()->orderBy('debut','desc')->get();
         $competences = $cv->competences()->orderBy('updated_at','desc')->get();
         $projets = $cv->projets()->orderBy('updated_at','desc')->get();
-
+        
         return Response()->json([
             'experiences' => $experiences,
             'formations'  => $formations,
             'competences' => $competences,
-            'projets'     =>  $projets
+            'projets'     =>  $projets,
             ]);
     }
